// src/types/emenda.ts
export interface Emenda {
  id: string;
  bloco: string | null;
  emenda: string | null;
  municipioBeneficiario: string | null;
  funcional: string | null;
  gnd: string | null;
  valorIndicado: number | null;
  objeto: string | null;
  alteracao: string | null;
  numeroProposta: string | null;
  valorEmpenhado: number | null;
  empenho: string | null;
  dataEmpenho: string | null;
  portariaConvenioContrato: string | null;
  valorAEmpenhar: number | null;
  pagamento: string | null;
  valorPago: number | null;
  valorASerPago: number | null;
  liderancas: string | null;
  createdAt: any;
  updatedAt: any;
}